mv AlignmentState to its own file
